+++ src/main/java/io/github/astrapi69/xml/XmlToObjectExtensions.java
@@ -24,15 +24,12 @@
  */
 package io.github.astrapi69.xml;
 
-import java.util.Map;
 import java.util.Objects;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.thoughtworks.xstream.XStream;
 
-import io.github.astrapi69.xml.factory.XStreamFactory;
 import io.github.astrapi69.xml.factory.XmlMapperFactory;
 
 /**
@@ -110,77 +107,6 @@ public final class XmlToObjectExtensions
 		Objects.requireNonNull(typeReference);
 		Objects.requireNonNull(xmlMapper);
 		return xmlMapper.readValue(xmlString, typeReference);
-	}
-
-	/**
-	 * Creates from the given xml string an java object.
-	 *
-	 * @param <T>
-	 *            the generic type of the return type
-	 * @param xmlString
-	 *            the xml as string object
-	 * @return the xml string
-	 */
-	public static <T> T toObjectWithXStream(final String xmlString)
-	{
-		return toObjectWithXStream(null, xmlString);
-	}
-
-	/**
-	 * Creates from the given xml string an Object. The given map hold the aliases. For more
-	 * information with aliasing see documation of xstream.
-	 *
-	 * @param <T>
-	 *            the generic type of the return type
-	 * @param xmlString
-	 *            the xml as string object
-	 * @param aliases
-	 *            the aliases
-	 * @return the created object from the given xml string.
-	 */
-	public static <T> T toObjectWithXStream(final String xmlString,
-		final Map<String, Class<?>> aliases)
-	{
-		return toObjectWithXStream(null, xmlString, aliases);
-	}
-
-	/**
-	 * Creates from the given xml string an java object.
-	 *
-	 * @param <T>
-	 *            the generic type of the return type
-	 * @param xstream
-	 *            the xstream object.
-	 * @param xmlString
-	 *            the xml as string object
-	 * @return the object
-	 */
-	public static <T> T toObjectWithXStream(final XStream xstream, final String xmlString)
-	{
-		return toObjectWithXStream(xstream, xmlString, null);
-	}
-
-	/**
-	 * Creates from the given xml string an java object. The given map hold the aliases. For more
-	 * information with aliasing see documentation of xstream.
-	 *
-	 * @param <T>
-	 *            the generic type of the return type
-	 * @param xstream
-	 *            the xstream object.
-	 * @param xmlString
-	 *            the xml
-	 * @param aliases
-	 *            the aliases
-	 * @return the object
-	 */
-	@SuppressWarnings("unchecked")
-	public static <T> T toObjectWithXStream(XStream xstream, final String xmlString,
-		final Map<String, Class<?>> aliases)
-	{
-		Objects.requireNonNull(xmlString);
-		xstream = XStreamFactory.initializeXStream(xstream, aliases);
-		return (T)xstream.fromXML(xmlString);
 	}
 
 }